src/requirements.txt
discord.py==1.7.3
numpy==1.21.4
opencv_python==4.5.4.58
pip==22.0.4
playsound==1.2.2
PyAutoGUI==0.9.53
pycryptodome==3.14.1
python-dotenv==0.20.0
pywin32==302
requests==2.27.1
scipy==1.8.0
sounddevice==0.4.4
WMI==1.5.1
